桥梁的本地数据

+++ app/src/main/java/itor/topnetwork/com/dxditor/hybrid/bean/bridge/BridgeBean.java
@@ -10,5 +10,5 @@ import java.util.List;
 public class BridgeBean {
     public List<String> xData;
     public List<Float> seriesData;
-    public int initValue;
+    public Float initValue;
 }
